[move]: move components to utilities
