+++ lib/src/utilities/components/greenfield_app_bar.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
-import '../utilities/design_system/app_texts.dart';
+import '../design_system/app_texts.dart';
 
 class GreenFieldAppBar extends StatelessWidget implements PreferredSizeWidget {
   final Color backgGroundColor;
