+++ lib/src/utilities/components/greenfield_list.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
-import '../utilities/design_system/app_texts.dart';
-import '../utilities/design_system/app_icons.dart';
+import '../design_system/app_icons.dart';
+import '../design_system/app_texts.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 
 class GreenFieldList extends StatelessWidget {
