+++ lib/src/views/main_view.dart
@@ -1,6 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import '../components/greenfield_tab_bar.dart';
+import '../utilities/components/greenfield_tab_bar.dart';
 
 class MainView extends StatelessWidget {
   final StatefulNavigationShell navigationShell;
