+++ lib/src/views/post/post_detail_view.dart
@@ -1,9 +1,9 @@
 import 'package:flutter/material.dart';
-import 'package:green_field/src/components/greenfield_app_bar.dart';
-import 'package:green_field/src/components/greenfield_comment_widget.dart';
-import 'package:green_field/src/components/greenfield_content_widget.dart';
-import 'package:green_field/src/components/greenfield_text_field.dart';
-import 'package:green_field/src/components/greenfield_user_info_widget.dart';
+import '../../utilities/components/greenfield_app_bar.dart';
+import '../../utilities/components/greenfield_comment_widget.dart';
+import '../../utilities/components/greenfield_content_widget.dart';
+import '../../utilities/components/greenfield_text_field.dart';
+import '../../utilities/components/greenfield_user_info_widget.dart';
 import 'package:green_field/src/enums/feature_type.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 import 'package:green_field/src/model/post.dart';
