+++ lib/src/views/recruitment/recruit_edit_view.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:green_field/src/components/greenfield_app_bar.dart';
-import 'package:green_field/src/components/greenfield_edit_section.dart'; // Assuming you have a GreenFieldEditSection
+import '../../utilities/components/greenfield_app_bar.dart';
+import '../../utilities/components/greenfield_edit_section.dart'; // Assuming you have a GreenFieldEditSection
 import 'package:green_field/src/enums/feature_type.dart';
 import 'package:green_field/src/utilities/extensions/theme_data_extension.dart';
 import '../../utilities/design_system/app_texts.dart';
